Add Log for user sending msg before the entry

+++ Bot/src/events/messageCreate.ts
@@ -1,16 +1,20 @@
 import { Client, Events, PermissionFlagsBits } from "discord.js";
-import EntryManager from "../app/Entry/EntryManager";
+
 import Entry from "../app/Entry/Entry";
 import { CHAN_DPLACE, ENTRY_CHANNELS } from "../config/config.guild";
+import threadLogger from "../utils/ThreadLogger";
 
 export default (client: Client) => {
-  client.on(Events.MessageCreate, (msg) => {
+  client.on(Events.MessageCreate, async (msg) => {
+    if (msg.author.bot) return;
+
     const channel = msg.channel;
     if (
       channel.id === CHAN_DPLACE &&
       !msg.member.permissions.has(PermissionFlagsBits.Administrator)
     ) {
-      msg.delete();
+      await threadLogger(msg);
+      await msg.delete();
     }
 
     const hasTempRole = Entry.getTempRole(msg.member) != null;
